my_app/requirements.txt
tkinter==0.0.0
requests==2.28.1
google-generativeai==0.0.1
PyPDF2==2.11.0